chore: got homepage to working state
I commented out the eye movement for the time being. I want to have the
site up and running by next Friday and want as few obstacles as
possible.

If we can get it to work by then, that would be awesome. I won't have
the ability to dedicate enough time to work on it and get everything
else done that the site needs, however.